Remove warnings when starting breeze
Breeze when started produced three warnings that were harmless,
but we should fix them to remove "false positives".
